Update website template
This commit focused on constructing the website according to my personal preference. I want my website to be a simple click and scroll. As such, alot of code was removed (especially the grid section), and new code was added, such as the layout for all five sections, and the interaction code, so that the viewer could use the navigation bar to go straight to their desired section.